monkeylearn_extract
Source: R/extractor.R

Access to Monkeylearn extractors modules

monkeylearn_extract(request, key = monkeylearn_key(quiet = TRUE),
  extractor_id = "ex_isnnZRbS", texts_per_req = 200, verbose = TRUE,
  params = NULL)

Arguments

request
A vector of characters (each text smaller than 50kB)

key
The API key

extractor_id
The ID of the extractor

texts_per_req
Number of texts to be fed through per request (max 200). Does not affect output, but may affect speed of processing.

verbose
whether to output messages about batch requests

params
parameters for the module as a named list. See the second example.

Value

A data.frame with the results whose attribute is a data.frame (tibble) "headers" including the number of remaining queries as "x.query.limit.remaining". Both data.frames include a column with the (list of) md5 checksum(s) of the corresponding text(s) computed using the digest digest function.

Details

Find IDs of extractors using https://app.monkeylearn.com/main/explore. Within the free plan, you can make up to 20 requests per minute.

You can use batch to send up to 200 texts to be analyzed within the API (classification or extraction) with each request. So for example, if you need to analyze 6000 tweets, instead of doing 6000 requests to the API, you can use batch to send 30 requests, each request with 200 tweets. The function automatically makes these batch calls and waits if there is a throttle limit error, but you might want to control the process yourself using several calls to the function.

You can check the number of calls you can still make in the API using attr(output, "headers")$x.query.limit.remaining and attr(output, "headers")$x.query.limit.limit.

Examples

# NOT RUN {
text <- "In the 19th century, the major European powers had gone to great lengths
to maintain a balance of power throughout Europe, resulting in the existence of
 a complex network of political and military alliances throughout the continent by 1900.[7]
  These had started in 1815, with the Holy Alliance between Prussia, Russia, and Austria.
  Then, in October 1873, German Chancellor Otto von Bismarck negotiated the League of
   the Three Emperors (German: Dreikaiserbund) between the monarchs of Austria-Hungary,
    Russia and Germany."
output <- monkeylearn_extract(request = text)
output
# example with parameters
text <- [1085 chars quoted with '"']
output <- monkeylearn_extract(text,
                              extractor_id = "ex_y7BPYzNG",
                              params = list(max_keywords = 3,
                                            use_company_names = 1))
attr(output, "headers")
# }